Example app update. README update. Dependency update.
