+++ examples/ShoutemAnimation/__tests__/App.js
@@ -1,0 +1,12 @@
+import 'react-native';
+import React from 'react';
+import App from '../App';
+
+// Note: test renderer must be required after react-native.
+import renderer from 'react-test-renderer';
+
+it('renders correctly', () => {
+  const tree = renderer.create(
+    <App />
+  );
+});
