+++ examples/ShoutemAnimation/app.json
@@ -1,0 +1,4 @@
+{
+  "name": "ShoutemAnimation",
+  "displayName": "ShoutemAnimation"
+}
