+++ examples/ShoutemAnimation/components/FadeInOut.js
