+++ examples/ShoutemAnimation/components/HeroHeader.js
