+++ examples/ShoutemAnimation/components/Parallax.js
