+++ examples/ShoutemAnimation/components/Rotate.js
