+++ examples/ShoutemAnimation/components/SlideInOut.js
