+++ examples/ShoutemAnimation/components/ZoomInOut.js
