+++ examples/ShoutemAnimation/components/ZoomInTouchable.js
