+++ examples/ShoutemAnimation/components/restaurants.json
