--- examples/ShoutemAnimation/index.android.js
@@ -1,22 +1,0 @@
-/**
- * Sample React Native App
- * https://github.com/facebook/react-native
- * @flow
- */
-
-import React, { Component } from 'react';
-import {
-  AppRegistry,
-} from 'react-native';
-
-import { App } from './App';
-
-class ShoutemAnimation extends Component {
-  render() {
-    return (
-      <App />
-    );
-  }
-}
-
-AppRegistry.registerComponent('ShoutemAnimation', () => ShoutemAnimation);
