+++ examples/ShoutemAnimation/ios/ShoutemAnimation.xcodeproj/xcshareddata/xcschemes/ShoutemAnimation-tvOS.xcscheme
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <Scheme
-   LastUpgradeVersion = "0620"
+   LastUpgradeVersion = "0820"
    version = "1.3">
    <BuildAction
       parallelizeBuildables = "NO"
@@ -14,9 +14,9 @@
             buildForAnalyzing = "YES">
             <BuildableReference
                BuildableIdentifier = "primary"
-               BlueprintIdentifier = "83CBBA2D1A601D0E00E9B192"
+               BlueprintIdentifier = "2D2A28121D9B038B00D4039D"
                BuildableName = "libReact.a"
-               BlueprintName = "React"
+               BlueprintName = "React-tvOS"
                ReferencedContainer = "container:../node_modules/react-native/React/React.xcodeproj">
             </BuildableReference>
          </BuildActionEntry>
@@ -28,9 +28,9 @@
             buildForAnalyzing = "YES">
             <BuildableReference
                BuildableIdentifier = "primary"
-               BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
-               BuildableName = "ShoutemAnimation.app"
-               BlueprintName = "ShoutemAnimation"
+               BlueprintIdentifier = "2D02E47A1E0B4A5D006451C7"
+               BuildableName = "ShoutemAnimation-tvOS.app"
+               BlueprintName = "ShoutemAnimation-tvOS"
                ReferencedContainer = "container:ShoutemAnimation.xcodeproj">
             </BuildableReference>
          </BuildActionEntry>
@@ -42,9 +42,9 @@
             buildForAnalyzing = "YES">
             <BuildableReference
                BuildableIdentifier = "primary"
-               BlueprintIdentifier = "00E356ED1AD99517003FC87E"
-               BuildableName = "ShoutemAnimationTests.xctest"
-               BlueprintName = "ShoutemAnimationTests"
+               BlueprintIdentifier = "2D02E48F1E0B4A5D006451C7"
+               BuildableName = "ShoutemAnimation-tvOSTests.xctest"
+               BlueprintName = "ShoutemAnimation-tvOSTests"
                ReferencedContainer = "container:ShoutemAnimation.xcodeproj">
             </BuildableReference>
          </BuildActionEntry>
@@ -60,9 +60,9 @@
             skipped = "NO">
             <BuildableReference
                BuildableIdentifier = "primary"
-               BlueprintIdentifier = "00E356ED1AD99517003FC87E"
-               BuildableName = "ShoutemAnimationTests.xctest"
-               BlueprintName = "ShoutemAnimationTests"
+               BlueprintIdentifier = "2D02E48F1E0B4A5D006451C7"
+               BuildableName = "ShoutemAnimation-tvOSTests.xctest"
+               BlueprintName = "ShoutemAnimation-tvOSTests"
                ReferencedContainer = "container:ShoutemAnimation.xcodeproj">
             </BuildableReference>
          </TestableReference>
@@ -70,9 +70,9 @@
       <MacroExpansion>
          <BuildableReference
             BuildableIdentifier = "primary"
-            BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
-            BuildableName = "ShoutemAnimation.app"
-            BlueprintName = "ShoutemAnimation"
+            BlueprintIdentifier = "2D02E47A1E0B4A5D006451C7"
+            BuildableName = "ShoutemAnimation-tvOS.app"
+            BlueprintName = "ShoutemAnimation-tvOS"
             ReferencedContainer = "container:ShoutemAnimation.xcodeproj">
          </BuildableReference>
       </MacroExpansion>
@@ -93,9 +93,9 @@
          runnableDebuggingMode = "0">
          <BuildableReference
             BuildableIdentifier = "primary"
-            BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
-            BuildableName = "ShoutemAnimation.app"
-            BlueprintName = "ShoutemAnimation"
+            BlueprintIdentifier = "2D02E47A1E0B4A5D006451C7"
+            BuildableName = "ShoutemAnimation-tvOS.app"
+            BlueprintName = "ShoutemAnimation-tvOS"
             ReferencedContainer = "container:ShoutemAnimation.xcodeproj">
          </BuildableReference>
       </BuildableProductRunnable>
@@ -112,9 +112,9 @@
          runnableDebuggingMode = "0">
          <BuildableReference
             BuildableIdentifier = "primary"
-            BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
-            BuildableName = "ShoutemAnimation.app"
-            BlueprintName = "ShoutemAnimation"
+            BlueprintIdentifier = "2D02E47A1E0B4A5D006451C7"
+            BuildableName = "ShoutemAnimation-tvOS.app"
+            BlueprintName = "ShoutemAnimation-tvOS"
             ReferencedContainer = "container:ShoutemAnimation.xcodeproj">
          </BuildableReference>
       </BuildableProductRunnable>
